Added Facade Layers And some of API Layers, Added Swagger, Some fixes in the services

+++ services/services-core/src/main/java/com.karen.shoppingbasket.services/StockMutationService.java
@@ -10,6 +10,6 @@ public interface StockMutationService {
 
     Long createStockMutation(Long productId, MutationType type, Integer count);
 
-    Long calculateStock(Long productId);
+    int calculateStock(Long productId);
 
 }
